Creating the FeedMapper file and setting the scope of the FeedImageDto and Root to be inside thee FeedMapper
